requirements.txt
Flask
gunicorn
Pillow
flask-cors
cloudinary
python-dotenv
